* Fixed: [MSVC] Ambiguous call to stdio's read. In modern C++ that's an overloaded function. So now TFileViewer::real_bufsize is of ssize_t type. I also changed some code to handle size_t return values. Reported by Eddie.

+++ tvision/include/tv/fileview.h
@@ -80,12 +80,12 @@ private:
 */
     TLineCollection *fileLines;
     char *buffer;
-    int real_bufsize;
+    ssize_t real_bufsize;
 
 protected:
 
     const char * const getBuffer() { return buffer; }
-    int getBufsize() { return real_bufsize; }
+    ssize_t getBufsize() { return real_bufsize; }
 
     int width;
     int count;
